Fix photo hash issue
Google changes the photo after a new one is loaded on the phone.
This messes up the hash that has been stored during a sync op. So,
a HashUpdateService updates the hashes tracked with new ones based
on the Google updated photos. Not ideal fix, but good enough for now.